Backend/vaalbackend/requirements.txt
django==4.2
djangorestframework==3.14.0
gunicorn==21.2.0
uvicorn==0.22.0
requests>=2.32.4
